src/app/pages/admin/aluno/dialog/confirm-dialog/confirm-dialog.component.ts
import { Component, Inject } from '@angular/core'
import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog'

@Component({
  selector: 'app-confirm-dialog',
  template: `
    <h1 mat-dialog-title>Confirmação</h1>
    <div mat-dialog-content>
      <p>{{ data.message }}</p>
    </div>
    <div mat-dialog-actions class="d-flex justify-content-between">
      <div class="flex-fill me-2">
        <button class="btn btn-success w-100" (click)="onConfirm()">Confirmar</button>
      </div>
      <div class="flex-fill">
        <button class="btn btn-danger w-100" (click)="onCancel()">Cancelar</button>
      </div>
    </div>
  `
})
export class ConfirmDialogComponent {
  constructor(
    public dialogRef: MatDialogRef<ConfirmDialogComponent>,
    @Inject(MAT_DIALOG_DATA) public data: { message: string }
  ) {}

  onCancel(): void {
    this.dialogRef.close(false)
  }

  onConfirm(): void {
    this.dialogRef.close(true)
  }
}
